Water: remove shallow pits
Add an additional parameter to `addWater` to remove basins/pits
with a depth less than the specified value

Reduces the number of basins and leaves only the larger lakes on
the terrain. Is visually more appealling then single cells of water.

Doesn't compensate for the "lost" water. This may be addressed
when soil humidity is introduced.